# openapi.yaml
openapi: 3.0.3
info:
  title: gdl-api
  description: gdl-api express api server + frontend + bots
  version: 0.2.31
  license:
    name: MIT
servers:
  - url: https://gdl.mcalec.dev/api
    description: production server
  - url: http://localhost:3030/api
    description: local server
...
paths:
...
  /uuid/{uuid}/{type}:
    get:
      tags:
        - Files
      summary: Retrieve file or directory information by UUID
      description: Retrieve detailed information about a file or directory using its UUID.
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
          description: The UUID of the file or directory to retrieve
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [file, directory]
          description: The type of the entry to retrieve (file or directory)
      responses:
        "200":
          description: File or directory information retrieved successfully
        "400":
          description: Invalid UUID or type parameter
        "404":
          description: File or directory not found
        "500":
          description: Internal server error
  /random:
    get:
      tags:
        - Random
      summary: Retrieve a random file
      description: Retrieve a random file from the database with metadata
      responses:
        "200":
          description: Random file metadata
        "500":
          description: Internal server error
  /random/:
    get:
      tags:
        - Random
      summary: Retrieve a random file (with trailing slash)
      description: Retrieve a random file from the database with metadata
      responses:
        "200":
          description: Random file metadata
        "500":
          description: Internal server error
  /random/image:
    get:
      tags:
        - Random
      summary: Retrieve a random image file
      description: Retrieve a random image file and stream it as a file download
      responses:
        "200":
          description: Image file binary content
        "500":
          description: Internal server error
  /random/image/:
    get:
      tags:
        - Random
      summary: Retrieve a random image file (with trailing slash)
      description: Retrieve a random image file and stream it as a file download
      responses:
        "200":
          description: Image file binary content
        "500":
          description: Internal server error
...
  /user/session/:
    get:
      tags:
        - User
      summary: Retrieve user sessions
      description: Get a list of all active sessions for the authenticated user
      responses:
        "200":
          description: List of user sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    uuid:
                      type: string
                    modified:
                      type: string
                      format: date-time
                    ip:
                      type: string
                    useragent:
                      type: string
        "401":
          description: User not authenticated
        "500":
          description: Internal server error
  /user/session/{uuid}:
    delete:
      tags:
        - User
      summary: Delete a specific user session
      description: Delete (logout) a specific session by UUID
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
          description: Session UUID to delete
      responses:
        "204":
          description: Session deleted successfully
        "400":
          description: Missing or invalid UUID
        "401":
          description: User not authenticated
        "500":
          description: Internal server error
...
  /admin/user/{id}:
    get:
      tags:
        - Admin
      summary: Get user information
      description: Retrieve detailed information about a specific user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      responses:
        "200":
          description: User information
    put:
      tags:
        - Admin
      summary: Update user information
      description: Modify user account details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      responses:
        "204":
          description: User updated successfully
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authenticated requests
security:
  - cookieAuth: []